Fix multiple counting.  Switch to allele-based instead of mutation-based.
I was double counting (and triple counting and occasionally 1000-counting) because leaf nodes returned by tree.leaves() are all the leaf nodes underneath that mutation, including on sub-branches where there have been additional mutations at that site.

So if a C changed to a G for the ancestor of 1000 people, then it changed back to a C for the ancestor of 400 of those people, then it changed back to a G for the ancestor of 100 of those people, I was counting 1000 people as having the first C>G mutation, 400 people as having the G>C mutation, and 100 people as having the second C>G mutation. I was counting 1000 people as 1500 people.

I have now switched to counting alleles instead of mutation events. So now, in the same situation, 700 people will be counted as C>G, 300 people won't be counted because they're back at the ancestral C state, and the date given will be that of the earliest C>G event at that site. This also means that mutations which were calculated to have occurred on separate trees will be lumped together, so a C>G that happened 50 years ago in Papua New Guinea will be grouped with a C>G that happened 100 years ago in France. I'm replacing genealogical relationships with genetic relationships.